Remove the JVM global compile strategy, switch default java compiler to zinc
On the assumption that Foursquare's switch to the isolated compile strategy will stick, this patch deletes the global strategy.

- Inline all of the isolated strategy into jvm_compile
- Merge pants.ini files
- Delete the global strategy and strategy interface
- Make `zinc` the default for java compiles since jmake does not support partitioned analysis
- Update `migrate_config` to indicate the global-related options will be ignored

Testing Done:
https://github.com/pantsbuild/pants/pull/2227

Reviewed at https://rbcommons.com/s/twitter/r/2852/
