+++ 3rdparty/BUILD
@@ -96,8 +96,13 @@ jar_library(name='slf4j-api',
 jar_library(name = 'spindle-runtime',
             jars = [
               jar(org = 'com.foursquare', name = 'spindle-runtime_2.10', rev = '3.0.0-M7'),
+            ],
+            dependencies = [
+              'contrib/spindle/3rdparty:rogue',
             ])
 
+# NB: we have two versions of thrift here due to scrooge requiring one version while apache
+# thrift requires another. this is not usually recommended
 jar_library(name='libthrift-0.9.2',
             jars = [
               jar(org='org.apache.thrift', name='libthrift', rev='0.9.2')
@@ -110,7 +115,17 @@ target(name='thrift-0.9.2',
          ':slf4j-api',
        ])
 
-target(name='thrift', dependencies = [ ':thrift-0.9.2' ])
+jar_library(name='libthrift-0.6.1',
+            jars = [
+              jar(org='org.apache.thrift', name='libthrift', rev='0.6.1')
+            ])
+
+target(name='thrift-0.6.1',
+       dependencies = [
+         ':commons-lang',
+         ':libthrift-0.6.1',
+         ':slf4j-api',
+       ])
 
 
 ###############
