+++ BUILD.tools
@@ -32,12 +32,12 @@ jar_library(name = 'scrooge-gen',
             jars = [
               jar(org='com.twitter', name='scrooge-generator_2.10', rev='3.20.0')
                 # scrooge requires libthrift 0.5.0-1 which is not available on
-                # the default maven repos. Force scrooge to use the lib thirft used
+                # the default maven repos. Force scrooge to use the lib thrift used
                 # by pants: 3rdparty:thrift
                 .exclude(org = 'org.apache.thrift', name = 'libthrift')
             ],
             dependencies = [
-              '3rdparty:thrift',
+              '3rdparty:thrift-0.6.1',
             ])
 
 jar_library(name = 'scrooge-linter',
@@ -46,5 +46,5 @@ jar_library(name = 'scrooge-linter',
                 .exclude(org = 'org.apache.thrift', name = 'libthrift')
             ],
             dependencies = [
-              '3rdparty:thrift',
+              '3rdparty:thrift-0.6.1',
             ])
