+++ pants.ini
@@ -74,9 +74,36 @@ ivy_settings: %(pants_supportdir)s/ivy/ivysettings.xml
 ivy_profile: %(pants_supportdir)s/ivy/ivy.xml
 
 
+[gen.scrooge]
+service_deps: {
+    'java': [
+      '//:scala-library',
+      '3rdparty:slf4j-api',
+      '3rdparty:thrift-0.6.1',
+      '3rdparty/jvm/com/twitter:finagle-thrift',
+      '3rdparty/jvm/com/twitter:scrooge-core',
+    ],
+    'scala': [
+      '3rdparty:thrift-0.6.1',
+      '3rdparty/jvm/com/twitter:finagle-thrift',
+      '3rdparty/jvm/com/twitter:scrooge-core',
+    ],
+  }
+structs_deps: {
+    'java': [
+      '3rdparty:thrift-0.6.1',
+      '3rdparty/jvm/com/twitter:scrooge-core',
+    ],
+    'scala': [
+      '3rdparty:thrift-0.6.1',
+      '3rdparty/jvm/com/twitter:scrooge-core',
+    ],
+  }
+
+
 [gen.thrift]
 gen_options: hashcode
-deps: ["3rdparty:thrift"]
+deps: ["3rdparty:thrift-0.9.2"]
 
 
 [compile.checkstyle]
@@ -116,14 +143,28 @@ skip: True
 [pycheck-except-statement]
 skip: True
 
-[compile.java]
-partition_size_hint: 1000000000
-compiler-bootstrap-tools: ["//:java-compiler"]
-jvm_options: ["-Xmx2G"]
-
 
 [compile.zinc]
-jvm_options: ["-Xmx2g", "-XX:MaxPermSize=256m", "-Dzinc.analysis.cache.limit=0"]
+worker_count: 4
+jvm_options: [
+    '-Xmx4g', '-XX:MaxPermSize=512m', '-XX:+UseConcMarkSweepGC', '-XX:ParallelGCThreads=4',
+    # bigger cache size for our big projects (default is just 5)
+    '-Dzinc.analysis.cache.limit=1000',
+  ]
+
+args: [
+    '-S-encoding', '-SUTF-8',
+    '-S-g:vars',
+  ]
+warning_args: [
+    '-S-deprecation',
+    '-S-unchecked',
+    # request warnings for http://www.scala-lang.org/api/2.10.4/index.html#scala.language$
+    '-S-feature',
+  ]
+no_warning_args: [
+    '-S-nowarn',
+  ]
 
 
 [jvm.run.jvm]
