--- pants.ini.isolated
@@ -1,33 +1,0 @@
-# An override config example that uses the `isolated` compile strategy for JVM languages.
-# ./pants --config-override=pants.ini.isolated <goal> <targets>
-
-[compile.apt]
-strategy: isolated
-
-
-[compile.java]
-use_jmake: False
-
-
-[compile.zinc]
-strategy: isolated
-worker_count: 4
-jvm_options: [
-    '-Xmx4g', '-XX:MaxPermSize=512m', '-XX:+UseConcMarkSweepGC', '-XX:ParallelGCThreads=4',
-    # bigger cache size for our big projects (default is just 5)
-    '-Dzinc.analysis.cache.limit=1000',
-  ]
-
-args: [
-    '-S-encoding', '-SUTF-8',
-    '-S-g:vars',
-  ]
-warning_args: [
-    '-S-deprecation',
-    '-S-unchecked',
-    # request warnings for http://www.scala-lang.org/api/2.10.4/index.html#scala.language$
-    '-S-feature',
-  ]
-no_warning_args: [
-    '-S-nowarn',
-  ]
