+++ src/python/pants/backend/codegen/tasks/BUILD
@@ -132,6 +132,7 @@ python_library(
     'src/python/pants/base:address',
     'src/python/pants/base:address_lookup_error',
     'src/python/pants/base:build_environment',
+    'src/python/pants/base:dep_lookup_error',
     'src/python/pants/base:workunit',
     'src/python/pants/util:memo',
   ],
