+++ src/python/pants/backend/jvm/subsystems/BUILD
@@ -5,7 +5,7 @@ python_library(
   name = 'jvm_tool_mixin',
   sources = ['jvm_tool_mixin.py'],
   dependencies = [
-    'src/python/pants/base:address_lookup_error',
+    'src/python/pants/base:dep_lookup_error',
     'src/python/pants/base:exceptions',
   ],
 )
