+++ src/python/pants/backend/jvm/tasks/jvm_compile/compile_context.py
@@ -18,11 +18,17 @@ class CompileContext(object):
   and a finalized compile in its permanent location.
   """
 
-  def __init__(self, target, analysis_file, classes_dir, sources):
+  def __init__(self, target, analysis_file, classes_dir, jar_file, sources):
     self.target = target
     self.analysis_file = analysis_file
     self.classes_dir = classes_dir
+    self.jar_file = jar_file
     self.sources = sources
+
+  @contextmanager
+  def open_jar(self, mode):
+    with open_zip(self.jar_file, mode=mode, compression=zipfile.ZIP_STORED) as jar:
+      yield jar
 
   @property
   def _id(self):
@@ -36,16 +42,3 @@ class CompileContext(object):
 
   def __hash__(self):
     return hash(self._id)
-
-
-class IsolatedCompileContext(CompileContext):
-  """Extends CompileContext to add a jar location."""
-
-  def __init__(self, target, analysis_file, classes_dir, jar_file, sources):
-    super(IsolatedCompileContext, self).__init__(target, analysis_file, classes_dir, sources)
-    self.jar_file = jar_file
-
-  @contextmanager
-  def open_jar(self, mode):
-    with open_zip(self.jar_file, mode=mode, compression=zipfile.ZIP_STORED) as jar:
-      yield jar
