+++ src/python/pants/backend/jvm/tasks/jvm_compile/java/java_compile.py
@@ -74,7 +74,7 @@ class JmakeCompile(JvmCompile):
   @classmethod
   def register_options(cls, register):
     super(JmakeCompile, cls).register_options(register)
-    register('--use-jmake', advanced=True, action='store_true', default=True,
+    register('--use-jmake', advanced=True, action='store_true', default=False,
              fingerprint=True,
              help='Use jmake to compile Java targets')
     cls.register_jvm_tool(register,
