+++ src/python/pants/backend/jvm/tasks/jvm_compile/jvm_compile.py
@@ -5,24 +5,49 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-import hashlib
+import functools
 import itertools
+import os
+import shutil
 import sys
-from collections import defaultdict
+from collections import OrderedDict, defaultdict
+from hashlib import sha1
 
 from pants.backend.core.tasks.group_task import GroupMember
 from pants.backend.jvm.subsystems.jvm_platform import JvmPlatform
 from pants.backend.jvm.targets.jar_library import JarLibrary
-from pants.backend.jvm.tasks.jvm_compile.jvm_compile_global_strategy import JvmCompileGlobalStrategy
-from pants.backend.jvm.tasks.jvm_compile.jvm_compile_isolated_strategy import \
-  JvmCompileIsolatedStrategy
+from pants.backend.jvm.tasks.classpath_util import ClasspathUtil
+from pants.backend.jvm.tasks.jvm_compile.compile_context import CompileContext
+from pants.backend.jvm.tasks.jvm_compile.execution_graph import (ExecutionFailure, ExecutionGraph,
+                                                                 Job)
+from pants.backend.jvm.tasks.jvm_compile.resource_mapping import ResourceMapping
 from pants.backend.jvm.tasks.nailgun_task import NailgunTaskBase
+from pants.base.build_environment import get_buildroot
 from pants.base.exceptions import TaskError
 from pants.base.fingerprint_strategy import TaskIdentityFingerprintStrategy
+from pants.base.worker_pool import Work, WorkerPool
 from pants.base.workunit import WorkUnitLabel
 from pants.goal.products import MultipleRootedProducts
 from pants.option.custom_types import list_option
 from pants.reporting.reporting_utils import items_to_report_element
+from pants.util.dirutil import fast_relpath, safe_mkdir, safe_rmtree, safe_walk
+from pants.util.fileutil import atomic_copy, create_size_estimators
+
+
+class CacheHitCallback(object):
+  """A serializable cache hit callback that cleans the class directory prior to cache extraction.
+
+  This class holds onto class directories rather than CompileContexts because CompileContexts
+  aren't picklable.
+  """
+
+  def __init__(self, cache_key_to_class_dir):
+    self._key_to_classes_dir = cache_key_to_class_dir
+
+  def __call__(self, cache_key):
+    class_dir = self._key_to_classes_dir.get(cache_key)
+    if class_dir:
+      safe_mkdir(class_dir, clean=True)
 
 
 class ResolvedJarAwareTaskIdentityFingerprintStrategy(TaskIdentityFingerprintStrategy):
@@ -61,14 +86,23 @@ class JvmCompile(NailgunTaskBase, GroupMember):
   mentioned below under "Subclasses must implement".
   """
 
+  size_estimators = create_size_estimators()
+
+  @classmethod
+  def size_estimator_by_name(cls, estimation_strategy_name):
+    return cls.size_estimators[estimation_strategy_name]
+
+  @staticmethod
+  def _analysis_for_target(analysis_dir, target):
+    return os.path.join(analysis_dir, target.id + '.analysis')
+
+  @staticmethod
+  def _portable_analysis_for_target(analysis_dir, target):
+    return JvmCompile._analysis_for_target(analysis_dir, target) + '.portable'
+
   @classmethod
   def register_options(cls, register):
     super(JvmCompile, cls).register_options(register)
-    register('--partition-size-hint', advanced=True, type=int, default=sys.maxint,
-             metavar='<# source files>',
-             help='Roughly how many source files to attempt to compile together. Set to a large '
-                  'number to compile all sources together. Set to 0 to compile target-by-target.')
-
     register('--jvm-options', advanced=True, type=list_option, default=[],
              help='Run the compiler with these JVM options.')
 
@@ -96,17 +130,20 @@ class JvmCompile(NailgunTaskBase, GroupMember):
              default=list(cls.get_no_warning_args_default()),
              help='Extra compiler args to use when warnings are disabled.')
 
-    register('--strategy', advanced=True, choices=['global', 'isolated'], default='global',
-             fingerprint=True,
-             help='Selects the compilation strategy to use. The "global" strategy uses a shared '
-                  'global classpath for all compiled classes, and the "isolated" strategy uses '
-                  'per-target classpaths.')
-
     register('--delete-scratch', advanced=True, default=True, action='store_true',
              help='Leave intermediate scratch files around, for debugging build problems.')
 
-    JvmCompileGlobalStrategy.register_options(register, cls._name, cls._supports_concurrent_execution)
-    JvmCompileIsolatedStrategy.register_options(register, cls._name, cls._supports_concurrent_execution)
+    register('--worker-count', advanced=True, type=int, default=1,
+             help='The number of concurrent workers to use when '
+                  'compiling with {task}.'.format(task=cls._name))
+
+    register('--size-estimator', advanced=True,
+             choices=list(cls.size_estimators.keys()), default='filesize',
+             help='The method of target size estimation.')
+
+    register('--capture-log', advanced=True, action='store_true', default=False,
+             fingerprint=True,
+             help='Capture compilation output to per-target logs.')
 
   @classmethod
   def product_types(cls):
@@ -175,7 +212,7 @@ class JvmCompile(NailgunTaskBase, GroupMember):
     return target.has_sources(self._file_suffix)
 
   def select_source(self, source_file_path):
-    """Source predicate for the strategy."""
+    """Source predicate for this task."""
     return source_file_path.endswith(self._file_suffix)
 
   def create_analysis_tools(self):
@@ -243,28 +280,67 @@ class JvmCompile(NailgunTaskBase, GroupMember):
     # The ivy confs for which we're building.
     self._confs = self.get_options().confs
 
-    # The compile strategy to use for analysis and classfile placement.
-    if self.get_options().strategy == 'global':
-      strategy_constructor = JvmCompileGlobalStrategy
-    else:
-      assert self.get_options().strategy == 'isolated'
-      strategy_constructor = JvmCompileIsolatedStrategy
-    self._strategy = strategy_constructor(self.context,
-                                          self.get_options(),
-                                          self.workdir,
-                                          self.create_analysis_tools(),
-                                          self._name,
-                                          self.select_source)
     # Maps CompileContext --> dict of upstream class to paths.
     self._upstream_class_to_paths = {}
+
+    # Mapping of relevant (as selected by the predicate) sources by target.
+    self._sources_by_target = None
+    self._sources_predicate = self.select_source
+
+    # Various working directories.
+    self._analysis_dir = os.path.join(self.workdir, 'isolated-analysis')
+    self._classes_dir = os.path.join(self.workdir, 'isolated-classes')
+    self._logs_dir = os.path.join(self.workdir, 'isolated-logs')
+    self._jars_dir = os.path.join(self.workdir, 'jars')
+
+    self._capture_log = self.get_options().capture_log
+    self._delete_scratch = self.get_options().delete_scratch
+    self._clear_invalid_analysis = self.get_options().clear_invalid_analysis
+
+    try:
+      worker_count = self.get_options().worker_count
+    except AttributeError:
+      # tasks that don't support concurrent execution have no worker_count registered
+      worker_count = 1
+    self._worker_count = worker_count
+
+    self._size_estimator = self.size_estimator_by_name(self.get_options().size_estimator)
+
+    self._worker_pool = None
+
+    self._analysis_tools = self.create_analysis_tools()
+
+  @property
+  def _analysis_parser(self):
+    return self._analysis_tools.parser
 
   def _fingerprint_strategy(self, classpath_products):
     return ResolvedJarAwareTaskIdentityFingerprintStrategy(self, classpath_products)
 
+  def ensure_analysis_tmpdir(self):
+    """Work in a tmpdir so we don't stomp the main analysis files on error.
+
+    A temporary, but well-known, dir in which to munge analysis/dependency files in before
+    caching. It must be well-known so we know where to find the files when we retrieve them from
+    the cache. The tmpdir is cleaned up in a shutdown hook, because background work
+    may need to access files we create there even after this method returns
+    :return: path of temporary analysis directory
+    """
+    analysis_tmpdir = os.path.join(self._workdir, 'analysis_tmpdir')
+    if self._delete_scratch:
+      self.context.background_worker_pool().add_shutdown_hook(
+        lambda: safe_rmtree(analysis_tmpdir))
+    safe_mkdir(analysis_tmpdir)
+    return analysis_tmpdir
+
   def pre_execute(self):
     # Only create these working dirs during execution phase, otherwise, they
     # would be wiped out by clean-all goal/task if it's specified.
-    self._strategy.pre_compile()
+    self.analysis_tmpdir = self.ensure_analysis_tmpdir()
+    safe_mkdir(self._analysis_dir)
+    safe_mkdir(self._classes_dir)
+    safe_mkdir(self._logs_dir)
+    safe_mkdir(self._jars_dir)
 
     # TODO(John Sirois): Ensuring requested product maps are available - if empty - should probably
     # be lifted to Task infra.
@@ -273,13 +349,43 @@ class JvmCompile(NailgunTaskBase, GroupMember):
     self._create_empty_products()
 
   def prepare_execute(self, chunks):
-    targets_in_chunks = list(itertools.chain(*chunks))
+    relevant_targets = list(itertools.chain(*chunks))
 
-    classpath_product = self.context.products.get_data('compile_classpath')
-    # Invoke the strategy's prepare_compile to prune analysis.
-    cache_manager = self.create_cache_manager(invalidate_dependents=True,
-                                              fingerprint_strategy=self._fingerprint_strategy(classpath_product))
-    self._strategy.prepare_compile(cache_manager, self.context.targets(), targets_in_chunks)
+    # Target -> sources (relative to buildroot).
+    # TODO(benjy): Should sources_by_target be available in all Tasks?
+    self._sources_by_target = self._compute_sources_by_target(relevant_targets)
+
+    # Update the classpath by adding relevant target's classes directories to its classpath.
+    compile_classpaths = self.context.products.get_data('compile_classpath')
+
+    with self.context.new_workunit('validate-{}-analysis'.format(self._name)):
+      for target in relevant_targets:
+        cc = self.compile_context(target)
+        safe_mkdir(cc.classes_dir)
+        compile_classpaths.add_for_target(target, [(conf, cc.classes_dir) for conf in self._confs])
+        self.validate_analysis(cc.analysis_file)
+
+    # This ensures the workunit for the worker pool is set
+    with self.context.new_workunit('isolation-{}-pool-bootstrap'.format(self._name)) \
+            as workunit:
+      # This uses workunit.parent as the WorkerPool's parent so that child workunits
+      # of different pools will show up in order in the html output. This way the current running
+      # workunit is on the bottom of the page rather than possibly in the middle.
+      self._worker_pool = WorkerPool(workunit.parent,
+                                     self.context.run_tracker,
+                                     self._worker_count)
+
+  def compile_context(self, target):
+    analysis_file = JvmCompile._analysis_for_target(self._analysis_dir, target)
+    classes_dir = os.path.join(self._classes_dir, target.id)
+    # Generate a short unique path for the jar to allow for shorter classpaths.
+    #   TODO: likely unnecessary after https://github.com/pantsbuild/pants/issues/1988
+    jar_file = os.path.join(self._jars_dir, '{}.jar'.format(sha1(target.id).hexdigest()[:12]))
+    return CompileContext(target,
+                          analysis_file,
+                          classes_dir,
+                          jar_file,
+                          self._sources_for_target(target))
 
   def execute_chunk(self, relevant_targets):
     if not relevant_targets:
@@ -289,7 +395,7 @@ class JvmCompile(NailgunTaskBase, GroupMember):
     fingerprint_strategy = self._fingerprint_strategy(classpath_product)
     # Invalidation check. Everything inside the with block must succeed for the
     # invalid targets to become valid.
-    partition_size_hint, locally_changed_targets = self._strategy.invalidation_hints(relevant_targets)
+    partition_size_hint, locally_changed_targets = (0, None)
     with self.invalidated(relevant_targets,
                           invalidate_dependents=True,
                           partition_size_hint=partition_size_hint,
@@ -303,30 +409,73 @@ class JvmCompile(NailgunTaskBase, GroupMember):
         # Register products for all the valid targets.
         # We register as we go, so dependency checking code can use this data.
         valid_targets = [vt.target for vt in invalidation_check.all_vts if vt.valid]
-        valid_compile_contexts = [self._strategy.compile_context(t) for t in valid_targets]
+        valid_compile_contexts = [self.compile_context(t) for t in valid_targets]
         self._register_vts(valid_compile_contexts)
 
-        # Invoke the strategy to execute compilations for invalid targets.
+        # Execute compilations for invalid targets.
         check_vts = (self.check_artifact_cache
             if self.artifact_cache_reads_enabled() else None)
         update_artifact_cache_vts_work = (self.get_update_artifact_cache_work
             if self.artifact_cache_writes_enabled() else None)
-        self._strategy.compile_chunk(invalidation_check,
-                                     self.context.targets(),
-                                     relevant_targets,
-                                     invalid_targets,
-                                     self.extra_compile_time_classpath_elements(),
-                                     check_vts,
-                                     self._compile_vts,
-                                     self._register_vts,
-                                     update_artifact_cache_vts_work)
+        self.compile_chunk(invalidation_check,
+                           self.context.targets(),
+                           relevant_targets,
+                           invalid_targets,
+                           self.extra_compile_time_classpath_elements(),
+                           check_vts,
+                           self._compile_vts,
+                           self._register_vts,
+                           update_artifact_cache_vts_work)
       else:
         # Nothing to build. Register products for all the targets in one go.
-        self._register_vts([self._strategy.compile_context(t) for t in relevant_targets])
+        self._register_vts([self.compile_context(t) for t in relevant_targets])
+
+  def compile_chunk(self,
+                    invalidation_check,
+                    all_targets,
+                    relevant_targets,
+                    invalid_targets,
+                    extra_compile_time_classpath_elements,
+                    check_vts,
+                    compile_vts,
+                    register_vts,
+                    update_artifact_cache_vts_work):
+    """Executes compilations for the invalid targets contained in a single chunk."""
+    assert invalid_targets, "compile_chunk should only be invoked if there are invalid targets."
+    # Get the classpath generated by upstream JVM tasks and our own prepare_compile().
+    compile_classpaths = self.context.products.get_data('compile_classpath')
+
+    extra_compile_time_classpath = self._compute_extra_classpath(
+        extra_compile_time_classpath_elements)
+
+    compile_contexts = self._create_compile_contexts_for_targets(all_targets)
+
+    # Now create compile jobs for each invalid target one by one.
+    jobs = self._create_compile_jobs(compile_classpaths,
+                                     compile_contexts,
+                                     extra_compile_time_classpath,
+                                     invalid_targets,
+                                     invalidation_check.invalid_vts_partitioned,
+                                     check_vts,
+                                     compile_vts,
+                                     register_vts,
+                                     update_artifact_cache_vts_work)
+
+    exec_graph = ExecutionGraph(jobs)
+    try:
+      exec_graph.execute(self._worker_pool, self.context.log)
+    except ExecutionFailure as e:
+      raise TaskError("Compilation failure: {}".format(e))
 
   def finalize_execute(self, chunks):
-    targets_in_chunks = list(itertools.chain(*chunks))
-    self._strategy.finalize_compile(targets_in_chunks)
+    targets = list(itertools.chain(*chunks))
+    # Replace the classpath entry for each target with its jar'd representation.
+    compile_classpaths = self.context.products.get_data('compile_classpath')
+    for target in targets:
+      cc = self.compile_context(target)
+      for conf in self._confs:
+        compile_classpaths.remove_for_target(target, [(conf, cc.classes_dir)])
+        compile_classpaths.add_for_target(target, [(conf, cc.jar_file)])
 
   def _compile_vts(self, vts, sources, analysis_file, upstream_analysis, classpath, outdir,
                    log_file, progress_message, settings):
@@ -365,10 +514,32 @@ class JvmCompile(NailgunTaskBase, GroupMember):
                      log_file, settings)
 
   def check_artifact_cache(self, vts):
-    post_process_cached_vts = lambda cvts: self._strategy.post_process_cached_vts(cvts)
-    cache_hit_callback=self._strategy.create_cache_hit_callback(vts)
-    return self.do_check_artifact_cache(vts, post_process_cached_vts=post_process_cached_vts,
+    post_process_cached_vts = lambda cvts: self.post_process_cached_vts(cvts)
+    cache_hit_callback = self.create_cache_hit_callback(vts)
+    return self.do_check_artifact_cache(vts,
+                                        post_process_cached_vts=post_process_cached_vts,
                                         cache_hit_callback=cache_hit_callback)
+
+  def create_cache_hit_callback(self, vts):
+    cache_key_to_classes_dir = {v.cache_key: self.compile_context(v.target).classes_dir
+                                for v in vts}
+    return CacheHitCallback(cache_key_to_classes_dir)
+
+  def post_process_cached_vts(self, cached_vts):
+    """Localizes the fetched analysis for targets we found in the cache.
+
+    This is the complement of `_write_to_artifact_cache`.
+    """
+    compile_contexts = []
+    for vt in cached_vts:
+      for target in vt.targets:
+        compile_contexts.append(self.compile_context(target))
+
+    for compile_context in compile_contexts:
+      portable_analysis_file = JvmCompile._portable_analysis_for_target(
+          self._analysis_dir, compile_context.target)
+      if os.path.exists(portable_analysis_file):
+        self._analysis_tools.localize(portable_analysis_file, compile_context.analysis_file)
 
   def _create_empty_products(self):
     make_products = lambda: defaultdict(MultipleRootedProducts)
@@ -384,6 +555,48 @@ class JvmCompile(NailgunTaskBase, GroupMember):
 
     self.context.products.safe_create_data('product_deps_by_src', dict)
 
+  def compute_classes_by_source(self, compile_contexts):
+    """Compute a map of (context->(src->classes)) for the given compile_contexts.
+
+    It's possible (although unfortunate) for multiple targets to own the same sources, hence
+    the top level division. Srcs are relative to buildroot. Classes are absolute paths.
+
+    Returning classes with 'None' as their src indicates that the compiler analysis indicated
+    that they were un-owned. This case is triggered when annotation processors generate
+    classes (or due to bugs in classfile tracking in zinc/jmake.)
+    """
+    buildroot = get_buildroot()
+    # Build a mapping of srcs to classes for each context.
+    classes_by_src_by_context = defaultdict(dict)
+    for compile_context in compile_contexts:
+      # Walk the context's jar to build a set of unclaimed classfiles.
+      unclaimed_classes = set()
+      with compile_context.open_jar(mode='r') as jar:
+        for name in jar.namelist():
+          if not name.endswith('/'):
+            unclaimed_classes.add(os.path.join(compile_context.classes_dir, name))
+
+      # Grab the analysis' view of which classfiles were generated.
+      classes_by_src = classes_by_src_by_context[compile_context]
+      if os.path.exists(compile_context.analysis_file):
+        products = self._analysis_parser.parse_products_from_path(compile_context.analysis_file,
+                                                                  compile_context.classes_dir)
+        for src, classes in products.items():
+          relsrc = os.path.relpath(src, buildroot)
+          classes_by_src[relsrc] = classes
+          unclaimed_classes.difference_update(classes)
+
+      # Any remaining classfiles were unclaimed by sources/analysis.
+      classes_by_src[None] = list(unclaimed_classes)
+    return classes_by_src_by_context
+
+  def class_name_for_class_file(self, compile_context, class_file_name):
+    if not class_file_name.endswith(".class"):
+      return None
+    assert class_file_name.startswith(compile_context.classes_dir)
+    class_file_name = class_file_name[len(compile_context.classes_dir) + 1:-len(".class")]
+    return class_file_name.replace("/", ".")
+
   def _register_vts(self, compile_contexts):
     classes_by_source = self.context.products.get_data('classes_by_source')
     classes_by_target = self.context.products.get_data('classes_by_target')
@@ -392,9 +605,9 @@ class JvmCompile(NailgunTaskBase, GroupMember):
     product_deps_by_src = self.context.products.get_data('product_deps_by_src')
 
     # Register class products (and resources generated by annotation processors.)
-    computed_classes_by_source_by_context = self._strategy.compute_classes_by_source(
+    computed_classes_by_source_by_context = self.compute_classes_by_source(
         compile_contexts)
-    resource_mapping = self._strategy.compute_resource_mapping(compile_contexts)
+    resource_mapping = ResourceMapping(self._classes_dir)
     for compile_context in compile_contexts:
       computed_classes_by_source = computed_classes_by_source_by_context[compile_context]
       target = compile_context.target
@@ -402,7 +615,7 @@ class JvmCompile(NailgunTaskBase, GroupMember):
 
       def add_products_by_target(files):
         for f in files:
-          clsname = self._strategy.class_name_for_class_file(compile_context, f)
+          clsname = self.class_name_for_class_file(compile_context, f)
           if clsname:
             # Is a class.
             classes_by_target[target].add_abs_paths(classes_dir, [f])
@@ -444,4 +657,251 @@ class JvmCompile(NailgunTaskBase, GroupMember):
       compile_classpath.add_for_target(compile_context.target, entries)
 
       if self.context.products.is_required_data('product_deps_by_src'):
-        product_deps_by_src[compile_context.target] = self._strategy.parse_deps(compile_context.analysis_file)
+        product_deps_by_src[compile_context.target] = \
+            self._analysis_parser.parse_deps_from_path(compile_context.analysis_file)
+
+  def _create_compile_contexts_for_targets(self, targets):
+    compile_contexts = OrderedDict()
+    for target in targets:
+      compile_context = self.compile_context(target)
+      compile_contexts[target] = compile_context
+    return compile_contexts
+
+  def _compute_classpath_entries(self, compile_classpaths,
+                                 target_closure,
+                                 compile_context,
+                                 extra_compile_time_classpath):
+    # Generate a classpath specific to this compile and target.
+    return ClasspathUtil.compute_classpath_for_target(compile_context.target, compile_classpaths,
+                                                      extra_compile_time_classpath, self._confs,
+                                                      target_closure)
+
+  def _upstream_analysis(self, compile_contexts, classpath_entries):
+    """Returns tuples of classes_dir->analysis_file for the closure of the target."""
+    # Reorganize the compile_contexts by class directory.
+    compile_contexts_by_directory = {}
+    for compile_context in compile_contexts.values():
+      compile_contexts_by_directory[compile_context.classes_dir] = compile_context
+    # If we have a compile context for the target, include it.
+    for entry in classpath_entries:
+      if not entry.endswith('.jar'):
+        compile_context = compile_contexts_by_directory.get(entry)
+        if not compile_context:
+          self.context.log.debug('Missing upstream analysis for {}'.format(entry))
+        else:
+          yield compile_context.classes_dir, compile_context.analysis_file
+
+  def _capture_log_file(self, target):
+    if self._capture_log:
+      return os.path.join(self._logs_dir, "{}.log".format(target.id))
+    return None
+
+  def exec_graph_key_for_target(self, compile_target):
+    return "compile({})".format(compile_target.address.spec)
+
+  def _create_compile_jobs(self, compile_classpaths, compile_contexts, extra_compile_time_classpath,
+                           invalid_targets, invalid_vts_partitioned, check_vts, compile_vts,
+                           register_vts, update_artifact_cache_vts_work):
+    def check_cache(vts):
+      """Manually checks the artifact cache (usually immediately before compilation.)
+
+      Returns true if the cache was hit successfully, indicating that no compilation is necessary.
+      """
+      if not check_vts:
+        return False
+      cached_vts, uncached_vts = check_vts([vts])
+      if not cached_vts:
+        self.context.log.debug('Missed cache during double check for {}'.format(vts.target.address.spec))
+        return False
+      assert cached_vts == [vts], (
+          'Cache returned unexpected target: {} vs {}'.format(cached_vts, [vts])
+      )
+      self.context.log.info('Hit cache during double check for {}'.format(vts.target.address.spec))
+      return True
+
+    def work_for_vts(vts, compile_context, target_closure):
+      progress_message = compile_context.target.address.spec
+      cp_entries = self._compute_classpath_entries(compile_classpaths,
+                                                   target_closure,
+                                                   compile_context,
+                                                   extra_compile_time_classpath)
+
+      upstream_analysis = dict(self._upstream_analysis(compile_contexts, cp_entries))
+
+      # Capture a compilation log if requested.
+      log_file = self._capture_log_file(compile_context.target)
+
+      # Double check the cache before beginning compilation
+      if not check_cache(vts):
+        # Mutate analysis within a temporary directory, and move it to the final location
+        # on success.
+        tmpdir = os.path.join(self.analysis_tmpdir, compile_context.target.id)
+        safe_mkdir(tmpdir)
+        tmp_analysis_file = self._analysis_for_target(
+            tmpdir, compile_context.target)
+        if os.path.exists(compile_context.analysis_file):
+          shutil.copy(compile_context.analysis_file, tmp_analysis_file)
+        target, = vts.targets
+        compile_vts(vts,
+                    compile_context.sources,
+                    tmp_analysis_file,
+                    upstream_analysis,
+                    cp_entries,
+                    compile_context.classes_dir,
+                    log_file,
+                    progress_message,
+                    target.platform)
+        atomic_copy(tmp_analysis_file, compile_context.analysis_file)
+
+        # Jar the compiled output.
+        self._create_context_jar(compile_context)
+
+      # Update the products with the latest classes.
+      register_vts([compile_context])
+
+      # Kick off the background artifact cache write.
+      if update_artifact_cache_vts_work:
+        self._write_to_artifact_cache(vts, compile_context, update_artifact_cache_vts_work)
+
+    jobs = []
+    invalid_target_set = set(invalid_targets)
+    for vts in invalid_vts_partitioned:
+      assert len(vts.targets) == 1, ("Requested one target per partition, got {}".format(vts))
+
+      # Invalidated targets are a subset of relevant targets: get the context for this one.
+      compile_target = vts.targets[0]
+      compile_context = compile_contexts[compile_target]
+      compile_target_closure = compile_target.closure()
+
+      # dependencies of the current target which are invalid for this chunk
+      invalid_dependencies = (compile_target_closure & invalid_target_set) - [compile_target]
+
+      jobs.append(Job(self.exec_graph_key_for_target(compile_target),
+                      functools.partial(work_for_vts, vts, compile_context, compile_target_closure),
+                      [self.exec_graph_key_for_target(target) for target in invalid_dependencies],
+                      self._size_estimator(compile_context.sources),
+                      # If compilation and analysis work succeeds, validate the vts.
+                      # Otherwise, fail it.
+                      on_success=vts.update,
+                      on_failure=vts.force_invalidate))
+    return jobs
+
+  def _create_context_jar(self, compile_context):
+    """Jar up the compile_context to its output jar location.
+
+    TODO(stuhood): In the medium term, we hope to add compiler support for this step, which would
+    allow the jars to be used as compile _inputs_ as well. Currently using jar'd compile outputs as
+    compile inputs would make the compiler's analysis useless.
+      see https://github.com/twitter-forks/sbt/tree/stuhood/output-jars
+    """
+    root = compile_context.classes_dir
+    with compile_context.open_jar(mode='w') as jar:
+      for abs_sub_dir, dirnames, filenames in safe_walk(root):
+        for name in dirnames + filenames:
+          abs_filename = os.path.join(abs_sub_dir, name)
+          arcname = fast_relpath(abs_filename, root)
+          jar.write(abs_filename, arcname)
+
+  def _write_to_artifact_cache(self, vts, compile_context, get_update_artifact_cache_work):
+    assert len(vts.targets) == 1
+    assert vts.targets[0] == compile_context.target
+
+    # Noop if the target is uncacheable.
+    if (compile_context.target.has_label('no_cache')):
+      return
+    vt = vts.versioned_targets[0]
+
+    # Set up args to relativize analysis in the background.
+    portable_analysis_file = self._portable_analysis_for_target(
+        self._analysis_dir, compile_context.target)
+    relativize_args_tuple = (compile_context.analysis_file, portable_analysis_file)
+
+    # Collect the artifacts for this target.
+    artifacts = []
+
+    def add_abs_products(p):
+      if p:
+        for _, paths in p.abs_paths():
+          artifacts.extend(paths)
+    # Resources.
+    resources_by_target = self.context.products.get_data('resources_by_target')
+    add_abs_products(resources_by_target.get(compile_context.target))
+    # Classes.
+    classes_by_target = self.context.products.get_data('classes_by_target')
+    add_abs_products(classes_by_target.get(compile_context.target))
+    # Log file.
+    log_file = self._capture_log_file(compile_context.target)
+    if log_file and os.path.exists(log_file):
+      artifacts.append(log_file)
+    # Jar.
+    artifacts.append(compile_context.jar_file)
+
+    # Get the 'work' that will publish these artifacts to the cache.
+    # NB: the portable analysis_file won't exist until we finish.
+    vts_artifactfiles_pair = (vt, artifacts + [portable_analysis_file])
+    update_artifact_cache_work = get_update_artifact_cache_work([vts_artifactfiles_pair])
+
+    # And execute it.
+    if update_artifact_cache_work:
+      work_chain = [
+          Work(self._analysis_tools.relativize, [relativize_args_tuple], 'relativize'),
+          update_artifact_cache_work
+      ]
+      self.context.submit_background_work_chain(work_chain, parent_workunit_name='cache')
+
+  def validate_analysis(self, path):
+    """Throws a TaskError for invalid analysis files."""
+    try:
+      self._analysis_parser.validate_analysis(path)
+    except Exception as e:
+      if self._clear_invalid_analysis:
+        self.context.log.warn("Invalid analysis detected at path {} ... pants will remove these "
+                              "automatically, but\nyou may experience spurious warnings until "
+                              "clean-all is executed.\n{}".format(path, e))
+        safe_delete(path)
+      else:
+        raise TaskError("An internal build directory contains invalid/mismatched analysis: please "
+                        "run `clean-all` if your tools versions changed recently:\n{}".format(e))
+
+  def _compute_sources_by_target(self, targets):
+    """Computes and returns a map target->sources (relative to buildroot)."""
+    def resolve_target_sources(target_sources):
+      resolved_sources = []
+      for target in target_sources:
+        if target.has_sources():
+          resolved_sources.extend(target.sources_relative_to_buildroot())
+      return resolved_sources
+
+    def calculate_sources(target):
+      sources = [s for s in target.sources_relative_to_buildroot() if self._sources_predicate(s)]
+      # TODO: Make this less hacky. Ideally target.java_sources will point to sources, not targets.
+      if hasattr(target, 'java_sources') and target.java_sources:
+        sources.extend(resolve_target_sources(target.java_sources))
+      return sources
+    return {t: calculate_sources(t) for t in targets}
+
+  def _sources_for_targets(self, targets):
+    """Returns a cached map of target->sources for the specified targets."""
+    if self._sources_by_target is None:
+      raise TaskError('self._sources_by_target not computed yet.')
+    return {t: self._sources_by_target.get(t, []) for t in targets}
+
+  def _sources_for_target(self, target):
+    """Returns the cached sources for the given target."""
+    if self._sources_by_target is None:
+      raise TaskError('self._sources_by_target not computed yet.')
+    return self._sources_by_target.get(target, [])
+
+  def _compute_extra_classpath(self, extra_compile_time_classpath_elements):
+    """Compute any extra compile-time-only classpath elements.
+
+    TODO(benjy): Model compile-time vs. runtime classpaths more explicitly.
+    TODO(benjy): Add a pre-execute goal for injecting deps into targets, so e.g.,
+    we can inject a dep on the scala runtime library and still have it ivy-resolve.
+    """
+    def extra_compile_classpath_iter():
+      for conf in self._confs:
+        for jar in extra_compile_time_classpath_elements:
+          yield (conf, jar)
+
+    return list(extra_compile_classpath_iter())
