+++ testprojects/tests/scala/org/pantsbuild/testproject/cp-directories/BUILD
@@ -1,0 +1,11 @@
+# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+junit_tests(
+  name = 'cp-directories',
+  dependencies = [
+    '3rdparty:junit',
+    '3rdparty:scalatest',
+  ],
+  sources=globs('*.scala'),
+)
