+++ tests/python/pants_test/backend/core/tasks/BUILD
@@ -30,7 +30,7 @@ python_tests(
   sources=['test_cache_cleanup.py'],
   coverage=['pants.backend.core.tasks.bash_completion'],
   dependencies=[
-    'src/python/pants/backend/jvm/tasks/jvm_compile:java',
+    'src/python/pants/backend/jvm/tasks/jvm_compile:zinc',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test:int-test',
     'src/python/pants/util:contextutil',
