+++ tests/python/pants_test/backend/core/tasks/test_cache_cleanup.py
@@ -7,7 +7,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 import os
 
-from pants.backend.jvm.tasks.jvm_compile.java.java_compile import JmakeCompile
+from pants.backend.jvm.tasks.jvm_compile.zinc.zinc_compile import ZincCompile
 from pants.util.contextutil import temporary_dir
 from pants.util.dirutil import touch
 from pants_test.pants_run_integration_test import PantsRunIntegrationTest
@@ -21,10 +21,10 @@ class CacheCleanupTest(PantsRunIntegrationTest):
             '--jvm-platform-default-platform=default']
 
   def test_leave_one(self):
-    """ Ensure that max-old of 1 removes all but one files"""
+    """Ensure that max-old of 1 removes all but one files"""
 
     with temporary_dir() as cache_dir:
-      artifact_dir = os.path.join(cache_dir, JmakeCompile.stable_name(),
+      artifact_dir = os.path.join(cache_dir, ZincCompile.stable_name(),
           'testprojects.src.java.org.pantsbuild.testproject.unicode.main.main')
 
       touch(os.path.join(artifact_dir, 'old_cache_test1'))
@@ -33,7 +33,7 @@ class CacheCleanupTest(PantsRunIntegrationTest):
       touch(os.path.join(artifact_dir, 'old_cache_test4'))
       touch(os.path.join(artifact_dir, 'old_cache_test5'))
 
-      config = {'cache.compile.java': {'write_to': [cache_dir]}}
+      config = {'cache.compile.zinc': {'write_to': [cache_dir]}}
 
       pants_run = self.run_pants(self.create_platform_args(6) +
                                  ['compile.java',
@@ -57,14 +57,14 @@ class CacheCleanupTest(PantsRunIntegrationTest):
       self.assertEqual(len(os.listdir(artifact_dir)), 1)
 
   def test_leave_none(self):
-    """ Ensure that max-old of zero removes all files
+    """Ensure that max-old of zero removes all files
 
     This test should ensure that conditional doesn't change to the simpler test of if max_old since
     we need to handle zero as well.
     """
 
     with temporary_dir() as cache_dir:
-      artifact_dir = os.path.join(cache_dir, JmakeCompile.stable_name(),
+      artifact_dir = os.path.join(cache_dir, ZincCompile.stable_name(),
           'testprojects.src.java.org.pantsbuild.testproject.unicode.main.main')
 
       touch(os.path.join(artifact_dir, 'old_cache_test1'))
@@ -73,7 +73,7 @@ class CacheCleanupTest(PantsRunIntegrationTest):
       touch(os.path.join(artifact_dir, 'old_cache_test4'))
       touch(os.path.join(artifact_dir, 'old_cache_test5'))
 
-      config = {'cache.compile.java': {'write_to': [cache_dir]}}
+      config = {'cache.compile.zinc': {'write_to': [cache_dir]}}
 
       pants_run = self.run_pants(self.create_platform_args(6) +
                                  ['compile.java',
