+++ tests/python/pants_test/backend/jvm/tasks/jvm_compile/java/BUILD
@@ -51,7 +51,7 @@ python_tests(
   name='java_compile_integration',
   sources=['test_java_compile_integration.py'],
   dependencies=[
-    'src/python/pants/backend/jvm/tasks/jvm_compile:java',
+    'src/python/pants/backend/jvm/tasks/jvm_compile:zinc',
     'src/python/pants/fs',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
