+++ tests/python/pants_test/backend/jvm/tasks/jvm_compile/zinc/test_zinc_compile_integration.py
@@ -9,38 +9,26 @@ import xml.etree.ElementTree as ET
 
 from pants.util.contextutil import open_tar, open_zip, temporary_dir
 from pants_test.backend.jvm.tasks.jvm_compile.base_compile_integration_test import BaseCompileIT
-from pants_test.testutils.compile_strategy_utils import provide_compile_strategies
 
 
-SHAPELESS_CLSNAME = 'ShapelessExample.class'
 SHAPELESS_CLSFILE = 'org/pantsbuild/testproject/unicode/shapeless/ShapelessExample.class'
 SHAPELESS_TARGET = 'testprojects/src/scala/org/pantsbuild/testproject/unicode/shapeless'
 
 
 class ZincCompileIntegrationTest(BaseCompileIT):
 
-  def test_scala_global_compile(self):
-    with self.do_test_compile(SHAPELESS_TARGET,
-                              'global',
-                              expected_files=[SHAPELESS_CLSNAME]) as found:
-
-      self.assertTrue(self.get_only(found, SHAPELESS_CLSNAME).endswith(SHAPELESS_CLSFILE))
-
-  def test_scala_isolated_compile_jar(self):
+  def test_scala_compile_jar(self):
     # NB: generated with:
     #   hashlib.sha1('testprojects.src.scala.org.pantsbuild.testproject.unicode.shapeless.shapeless').hexdigest()[:12]
     jar_suffix = 'fd9f49e1153b.jar'
     with self.do_test_compile(SHAPELESS_TARGET,
-                              'isolated',
                               expected_files=[jar_suffix]) as found:
       with open_zip(self.get_only(found, jar_suffix), 'r') as jar:
         self.assertTrue(jar.getinfo(SHAPELESS_CLSFILE),
                         'Expected a jar containing the expected class.')
 
-  @provide_compile_strategies
-  def test_scala_empty_compile(self, strategy):
+  def test_scala_empty_compile(self):
     with self.do_test_compile('testprojects/src/scala/org/pantsbuild/testproject/emptyscala',
-                              strategy,
                               expected_files=[]) as found:
       # no classes generated by this target
       pass
@@ -48,35 +36,25 @@ class ZincCompileIntegrationTest(BaseCompileIT):
   def test_scala_shared_sources(self):
     clsname = 'SharedSources.class'
 
-    def test(strategy, count):
-      with self.do_test_compile('testprojects/src/scala/org/pantsbuild/testproject/sharedsources::',
-                                strategy,
-                                expected_files=[clsname]) as found:
-        classes = found[clsname]
-        self.assertEqual(count, len(classes))
-        for cls in classes:
-          self.assertTrue(cls.endswith(
-            'org/pantsbuild/testproject/sharedsources/SharedSources.class'))
+    with self.do_test_compile('testprojects/src/scala/org/pantsbuild/testproject/sharedsources::',
+                              expected_files=[clsname]) as found:
+      classes = found[clsname]
+      self.assertEqual(2, len(classes))
+      for cls in classes:
+        self.assertTrue(cls.endswith(
+          'org/pantsbuild/testproject/sharedsources/SharedSources.class'))
 
-    # We expect a single output class for the global strategy.
-    test('global', 1)
-    # But the isolated strategy should result in a class per target.
-    test('isolated', 2)
-
-  def test_scala_isolated_failure(self):
+  def test_scala_failure(self):
     """With no initial analysis, a failed compilation shouldn't leave anything behind."""
     analysis_file = 'testprojects.src.scala.' \
         'org.pantsbuild.testproject.compilation_failure.compilation_failure.analysis'
     with self.do_test_compile('testprojects/src/scala/org/pantsbuild/testprojects/compilation_failure',
-                              'isolated',
                               expected_files=[analysis_file],
                               expect_failure=True) as found:
       self.assertEqual(0, len(found[analysis_file]))
 
-  @provide_compile_strategies
-  def test_scala_with_java_sources_compile(self, strategy):
+  def test_scala_with_java_sources_compile(self):
     with self.do_test_compile('testprojects/src/scala/org/pantsbuild/testproject/javasources',
-                              strategy,
                               expected_files=['ScalaWithJavaSources.class',
                                               'JavaSource.class']) as found:
 
@@ -88,10 +66,8 @@ class ZincCompileIntegrationTest(BaseCompileIT):
           self.get_only(found, 'JavaSource.class').endswith(
               'org/pantsbuild/testproject/javasources/JavaSource.class'))
 
-  @provide_compile_strategies
-  def test_scalac_plugin_compile(self, strategy):
+  def test_scalac_plugin_compile(self):
     with self.do_test_compile('testprojects/src/scala/org/pantsbuild/testproject/scalac/plugin',
-                              strategy,
                               expected_files=['HelloScalac.class', 'scalac-plugin.xml']) as found:
 
       self.assertTrue(
@@ -113,7 +89,6 @@ class ZincCompileIntegrationTest(BaseCompileIT):
             workdir,
             cachedir,
             'testprojects/src/scala/org/pantsbuild/testproject/emptyscala',
-            'isolated',
             extra_args=[
               '--compile-zinc-args=-recompile-all-fraction',
               '--compile-zinc-args=0.5',
